Markov Chain Classifier

+++ archmm/estimation/clustering.pxd
@@ -5,7 +5,7 @@ cimport numpy as cnp
 
 cdef class ClusterSet:
     cdef cnp.double_t[:, :] data
-    cdef Py_ssize_t[:, :] clusters
+    cdef int[:, :] clusters
     cdef size_t* cluster_sizes
     cdef size_t n_clusters, n_points, point_dim
